Move test middlewares
