+++ src/Mediator.Net.Test/Cancellation/CancellationFixture.cs
@@ -4,11 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using Mediator.Net.Binding;
-using Mediator.Net.Test.Middlewares;
 using Mediator.Net.TestUtil.Handlers.CommandHandlers;
 using Mediator.Net.TestUtil.Handlers.EventHandlers;
 using Mediator.Net.TestUtil.Handlers.RequestHandlers;
 using Mediator.Net.TestUtil.Messages;
+using Mediator.Net.TestUtil.Middlewares;
 using Mediator.Net.TestUtil.TestUtils;
 using Shouldly;
 using Xunit;
