+++ src/Mediator.Net.Test/TestPerformance/SameMediatorSend1000CommandsOneMiddlewareEachPipe.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading.Tasks;
 using Mediator.Net.Binding;
-using Mediator.Net.Test.Middlewares;
 using Mediator.Net.TestUtil.Handlers.CommandHandlers;
 using Mediator.Net.TestUtil.Messages;
 using Mediator.Net.TestUtil.TestUtils;
+using Mediator.Net.TestUtil.Middlewares;
 using Shouldly;
 using TestStack.BDDfy;
 using Xunit;
