+++ src/Mediator.Net.Test/TestPipeline/ResponseShouldPropagateAcrossManyMiddlewares.cs
@@ -1,11 +1,11 @@
 ﻿using Mediator.Net.Binding;
-using Mediator.Net.Test.Middlewares;
 using Shouldly;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Mediator.Net.TestUtil.Handlers.RequestHandlers;
 using Mediator.Net.TestUtil.Messages;
+using Mediator.Net.TestUtil.Middlewares;
 using TestStack.BDDfy;
 using Xunit;
