+++ src/Mediator.Net.Test/TestPipeline/UselessMiddlewareShouldNotBeExecuted.cs
@@ -2,13 +2,13 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Mediator.Net.Binding;
-using Mediator.Net.Test.Middlewares;
-using Mediator.Net.TestUtil.Handlers.CommandHandlers;
 using Mediator.Net.TestUtil.Messages;
+using Mediator.Net.TestUtil.Middlewares;
 using Mediator.Net.TestUtil.TestUtils;
 using Shouldly;
 using TestStack.BDDfy;
 using Xunit;
+using Mediator.Net.TestUtil.Handlers.CommandHandlers;
 
 namespace Mediator.Net.Test.TestPipeline
 {
@@ -29,8 +29,6 @@ namespace Mediator.Net.Test.TestPipeline
                 x.UseUselessMiddleware();
             })
             .Build();
-            
-
         }
 
         async Task WhenACommandIsSent()
