+++ src/Mediator.Net.TestUtil/Handlers/EventHandlers/TestEventHandler.cs
@@ -18,4 +18,15 @@ namespace Mediator.Net.TestUtil.Handlers.EventHandlers
             return Task.FromResult(0);
         }
     }
+
+    public class MultiTestEventHandlerHandleTheSameEvent : IEventHandler<TestEvent>
+    {
+        public Task Handle(IReceiveContext<TestEvent> context, CancellationToken cancellationToken)
+        {
+            TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
+            RubishBox.Rublish.Add(nameof(MultiTestEventHandlerHandleTheSameEvent));
+            Console.WriteLine($"Hi, i am event {context.Message.Id}");
+            return Task.FromResult(0);
+        }
+    }
 }
