--- src/Mediator.Net.TestUtil/Handlers/EventHandlers/TestEventHandler2.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Mediator.Net.Context;
-using Mediator.Net.Contracts;
-using Mediator.Net.TestUtil.Messages;
-using Mediator.Net.TestUtil.TestUtils;
-
-namespace Mediator.Net.TestUtil.Handlers.EventHandlers
-{
-    public class TestEventHandler2 : IEventHandler<TestEvent>
-    {
-        public Task Handle(IReceiveContext<TestEvent> context, CancellationToken cancellationToken)
-        {
-            RubishBox.Rublish.Add(nameof(TestEventHandler2));
-            Console.WriteLine($"Hi, i am number 2 event {context.Message.Id}");
-            return Task.FromResult(0);
-        }
-    }
-}
