+++ src/Mediator.Net.TestUtil/Middlewares/DummySave.cs
@@ -4,19 +4,20 @@ using System.Threading.Tasks;
 using Mediator.Net.Context;
 using Mediator.Net.Contracts;
 using Mediator.Net.Pipeline;
+using Mediator.Net.TestUtil.TestUtils;
 
-namespace Mediator.Net.Test.Middlewares
+namespace Mediator.Net.TestUtil.Middlewares
 {
-    public static class SecurityInfo
+    public static class DummySave
     {
-        public static void UseSecurityInfo<TContext>(this IPipeConfigurator<TContext> configurator)
+        public static void UseDummySave<TContext>(this IPipeConfigurator<TContext> configurator)
             where TContext : IContext<IMessage>
         {
-            configurator.AddPipeSpecification(new SecurityInfoSpecification<TContext>());
+            configurator.AddPipeSpecification(new DummySaveSpecification<TContext>());
         }
     }
 
-    class SecurityInfoSpecification<TContext> : IPipeSpecification<TContext>
+    public class DummySaveSpecification<TContext> : IPipeSpecification<TContext> 
         where TContext : IContext<IMessage>
     {
         public bool ShouldExecute(TContext context, CancellationToken cancellationToken)
@@ -26,17 +27,24 @@ namespace Mediator.Net.Test.Middlewares
 
         public Task ExecuteBeforeConnect(TContext context, CancellationToken cancellationToken)
         {
-            context.MetaData["UserName"] = "hello";
+            TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
             return Task.FromResult(0);
         }
 
         public Task Execute(TContext context, CancellationToken cancellationToken)
         {
+            if (ShouldExecute(context, cancellationToken))
+            {
+                TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
+                RubishBox.Rublish.Add(nameof(DummySave.UseDummySave));
+            }
             return Task.FromResult(0);
         }
 
         public Task ExecuteAfterConnect(TContext context, CancellationToken cancellationToken)
         {
+            if (ShouldExecute(context, cancellationToken))
+                TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
             return Task.FromResult(0);
         }
