+++ src/Mediator.Net.TestUtil/Middlewares/MiddlewareThrowExBeforeConnect.cs
@@ -4,19 +4,20 @@ using System.Threading.Tasks;
 using Mediator.Net.Context;
 using Mediator.Net.Contracts;
 using Mediator.Net.Pipeline;
+using Mediator.Net.TestUtil.TestUtils;
 
-namespace Mediator.Net.Test.Middlewares
+namespace Mediator.Net.TestUtil.Middlewares
 {
-    public static class SecurityInfo2
+    public static class MiddlewareThrowExBeforeConnect
     {
-        public static void UseSecurityInfo2<TContext>(this IPipeConfigurator<TContext> configurator)
+        public static void UseMiddlewareThrowExBeforeConnect<TContext>(this IPipeConfigurator<TContext> configurator)
             where TContext : IContext<IMessage>
         {
-            configurator.AddPipeSpecification(new SecurityInfoSpecification2<TContext>());
+            configurator.AddPipeSpecification(new MiddlewareThrowExBeforeConnectSpecification<TContext>());
         }
     }
 
-    class SecurityInfoSpecification2<TContext> : IPipeSpecification<TContext>
+    public class MiddlewareThrowExBeforeConnectSpecification<TContext> : IPipeSpecification<TContext> 
         where TContext : IContext<IMessage>
     {
         public bool ShouldExecute(TContext context, CancellationToken cancellationToken)
@@ -26,23 +27,24 @@ namespace Mediator.Net.Test.Middlewares
 
         public Task ExecuteBeforeConnect(TContext context, CancellationToken cancellationToken)
         {
-            return Task.FromResult(0);
+            throw new Exception();
         }
 
         public Task Execute(TContext context, CancellationToken cancellationToken)
         {
-            if (ShouldExecute(context, cancellationToken))
-                context.MetaData["Password"] = "password";
             return Task.FromResult(0);
         }
 
         public Task ExecuteAfterConnect(TContext context, CancellationToken cancellationToken)
         {
+            if (ShouldExecute(context, cancellationToken))
+                Console.WriteLine($"After 1: {context.Message}");
             return Task.FromResult(0);
         }
 
         public void OnException(Exception ex, TContext context)
         {
+            RubishBox.Rublish.Add(ex);
             throw ex;
         }
     }
