+++ src/Mediator.Net.TestUtil/Middlewares/NoWorkMiddleware.cs
@@ -4,20 +4,19 @@ using System.Threading.Tasks;
 using Mediator.Net.Context;
 using Mediator.Net.Contracts;
 using Mediator.Net.Pipeline;
-using Mediator.Net.TestUtil.TestUtils;
 
-namespace Mediator.Net.Test.Middlewares
+namespace Mediator.Net.TestUtil.Middlewares
 {
-    static class ConsoleLog2
+    public static class NoWorkMiddleware
     {
-        public static void UseConsoleLogger2<TContext>(this IPipeConfigurator<TContext> configurator)
+        public static void UseNoWorkMiddleware<TContext>(this IPipeConfigurator<TContext> configurator)
             where TContext : IContext<IMessage>
         {
-            configurator.AddPipeSpecification(new ConsoleLogSpecification2<TContext>());
+            configurator.AddPipeSpecification(new NoWorkMiddlewareSpecification<TContext>());
         }
     }
 
-    class ConsoleLogSpecification2<TContext> : IPipeSpecification<TContext> 
+    public class NoWorkMiddlewareSpecification<TContext> : IPipeSpecification<TContext> 
         where TContext : IContext<IMessage>
     {
         public bool ShouldExecute(TContext context, CancellationToken cancellationToken)
@@ -27,30 +26,21 @@ namespace Mediator.Net.Test.Middlewares
 
         public Task ExecuteBeforeConnect(TContext context, CancellationToken cancellationToken)
         {
-            TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
             return Task.FromResult(0);
         }
 
         public Task Execute(TContext context, CancellationToken cancellationToken)
         {
-            if (ShouldExecute(context, cancellationToken))
-            {
-                TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
-                RubishBox.Rublish.Add(nameof(ConsoleLog2.UseConsoleLogger2));
-            }
             return Task.FromResult(0);
         }
 
         public Task ExecuteAfterConnect(TContext context, CancellationToken cancellationToken)
         {
-            if (ShouldExecute(context, cancellationToken))
-                TokenRecorder.Recorder.Add(cancellationToken.GetHashCode());
             return Task.FromResult(0);
         }
 
         public void OnException(Exception ex, TContext context)
         {
-            RubishBox.Rublish.Add(ex);
             throw ex;
         }
     }
